backend/pyproject.toml
[tool.poetry]
name = "backend-template"
version = "1.0.0"
description = "backend_template"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
werkzeug = "^2.3.6"
pytest = "^7.4.0"
redis = "^4.6.0"
python-jose = "^3.3.0"
pycryptodome = "^3.18.0"
requests = "^2.31.0"
python-multipart = "^0.0.6"
sqlalchemy-utils = "^0.41.1"
cryptography = "^41.0.3"
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.20"
httpx = "^0.25.0"
prometheus-fastapi-instrumentator = "^6.1.0"
python-json-logger = "^2.0.7"
logfire = {extras = ["fastapi"], version = "^0.46.1"}
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
